+++ feincms/contrib/preview/urls.py
@@ -1,8 +1,11 @@
-from django.conf.urls import url
+try:
+    from django.urls import re_path
+except ImportError:
+    from django.conf.urls import url as re_path
 
 from feincms.contrib.preview.views import PreviewHandler
 
 
 urlpatterns = [
-    url(r"^(.*)/_preview/(\d+)/$", PreviewHandler.as_view(), name="feincms_preview")
+    re_path(r"^(.*)/_preview/(\d+)/$", PreviewHandler.as_view(), name="feincms_preview")
 ]