+++ feincms/management/commands/medialibrary_orphans.py
@@ -3,8 +3,8 @@ from __future__ import absolute_import, unicode_literals
 import os
 
 from django.core.management.base import NoArgsCommand
-from django.utils.encoding import force_text
 
+from feincms._internal import force_text
 from feincms.module.medialibrary.models import MediaFile