+++ feincms/signals.py
@@ -16,6 +16,6 @@ from django.dispatch import Signal
 # This signal is sent when an item editor managed object is completely
 # saved, especially including all foreign or manytomany dependencies.
 
-itemeditor_post_save_related = Signal(providing_args=["instance", "created"])
+itemeditor_post_save_related = Signal()
 
 # ------------------------------------------------------------------------